+++ src/main/java/com/katok09/realestate/management/service/UserDetailsServiceImpl.java
@@ -139,7 +139,7 @@ public class UserDetailsServiceImpl implements UserDetailsService {
     /**
      * ユーザーIDを取得
      */
-    public Long getUserId() {
+    public int getUserId() {
       return user.getId();
     }
